refactor(web): extract shared SelectSkeleton and address review feedback
Extract duplicated SelectSkeleton from ModelSelect and RevisionSelect
into a shared component. Use isDisabled (not disabled) to gate chevron
visibility for consistency. Drop unnecessary waitFor in a synchronous
ModelSelect test and switch RevisionSelect tests to it() to match the
established convention.
